Moved input handler to controller
