+++ core/src/com/group66/game/screens/OnePlayerGameScreen.java
@@ -1,13 +1,10 @@
 package com.group66.game.screens;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.group66.game.BustaMove;
 import com.group66.game.cannon.BallAnimationLoader;
-import com.group66.game.helpers.AudioManager;
-import com.group66.game.input.InputHandler;
 import com.group66.game.screencontrollers.OnePlayerGameController;
 import com.group66.game.screencontrollers.AbstractGameController.GameState;
 import com.group66.game.settings.Config;
@@ -26,10 +23,8 @@ public class OnePlayerGameScreen extends AbstractGameScreen {
      */
     public OnePlayerGameScreen(Boolean randomLevel, int level) {
         gameController = new OnePlayerGameController(randomLevel, level);
-        inputHandler = new InputHandler();
         
         BustaMove.getGameInstance().getDynamicSettings().setRandomLevel(randomLevel, true);
-        setup_keys();
         loadRelatedGraphics();
         BallAnimationLoader.load();
     }
@@ -63,9 +58,6 @@ public class OnePlayerGameScreen extends AbstractGameScreen {
     @Override
     public void render(float delta) {
         try {
-            /* Handle input keys */
-            inputHandler.run();
-    
             //The game is about to end, dispose of the assets already
             if (gameController.getGameManager1().isGameEnded()) {
                 dispose();
@@ -93,72 +85,5 @@ public class OnePlayerGameScreen extends AbstractGameScreen {
         } catch (Exception e) {
             e.printStackTrace();
         }
-    }
-
-    /**
-     * Setup the keys used in the game screen keys.
-     */
-    private void setup_keys() {
-        // Setup the game keys
-        inputHandler.registerKeyMap("Shoot", Keys.SPACE);
-        inputHandler.registerKeyMap("Shoot", Keys.BACKSPACE);
-        inputHandler.registerKeyMap("Shoot", Keys.W);
-        inputHandler.registerKeyMap("Shoot", Keys.UP);
-        inputHandler.registerKeyMap("Aim Left", Keys.A);
-        inputHandler.registerKeyMap("Aim Left", Keys.LEFT);
-        inputHandler.registerKeyMap("Aim Right", Keys.D);
-        inputHandler.registerKeyMap("Aim Right", Keys.RIGHT);
-        inputHandler.registerKeyMap("Place Ball", Keys.ENTER);
-        inputHandler.registerKeyMap("Toggle Pause", Keys.ESCAPE);
-        inputHandler.registerKeyMap("Toggle mute", Keys.M);
-
-        /* Register key names to functions */
-        inputHandler.registerKeyPressedFunc("Aim Left",
-                new InputHandler.KeyCommand() {
-                    public void runCommand() {
-                        try {
-                            gameController.getGameManager1().cannon.cannonAimAdjust(Config.CANNON_AIM_DELTA);
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
-                    }
-            });
-
-        inputHandler.registerKeyPressedFunc("Aim Right",
-                new InputHandler.KeyCommand() {
-                    public void runCommand() {
-                        try {
-                            gameController.getGameManager1().cannon.cannonAimAdjust(-Config.CANNON_AIM_DELTA);
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
-                    }
-            });
-
-        inputHandler.registerKeyJustPressedFunc("Shoot",
-                new InputHandler.KeyCommand() {
-                    public void runCommand() {
-                        try {
-                            gameController.getGameManager1().shootBall();
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
-                    }
-            });
-
-        inputHandler.registerKeyJustPressedFunc("Toggle Pause",
-                new InputHandler.KeyCommand() {
-                    public void runCommand() {
-                        gameController.togglePause();
-                    }
-            });
-
-        inputHandler.registerKeyJustPressedFunc("Toggle mute",
-                new InputHandler.KeyCommand() {
-                    public void runCommand() {
-                        AudioManager.toggleMute();
-                    }
-            });
-
     }
 }
